requirements.txt
pytest~=9.0.2
Django~=5.0.6
djangorestframework~=3.15.2
ruff~=0.14.13